// frontend/src/types/generated/realtime.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export type paths = Record<string, never>;
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /** MediaAnswer */
        MediaAnswer: {
            /** Client Session Id */
            client_session_id: string;
            /** Connection Id */
            connection_id: string;
            /** Sdp */
            sdp: string;
            /**
             * Type
             * @default answer
             * @constant
             */
            type?: "answer";
        };
        /** MediaControl */
        MediaControl: {
            /**
             * Action
             * @enum {string}
             */
            action: "ping" | "interrupt" | "finish" | "commit" | "discard" | "ack" | "play_question";
            /** Connection Id */
            connection_id: string;
            /**
             * Playback Id
             * @default null
             */
            playback_id?: string | null;
            /**
             * Request Id
             * @default null
             */
            request_id?: string | null;
            /**
             * Samples
             * @default null
             */
            samples?: number | null;
        };
        /** MediaOffer */
        MediaOffer: {
            /**
             * Client Session Id
             * Format: uuid
             */
            client_session_id: string;
            /** Sdp */
            sdp: string;
            /**
             * Type
             * @default offer
             * @constant
             */
            type?: "offer";
        };
        /** MediaPlaybackReport */
        MediaPlaybackReport: {
            /** Client Reported Samples */
            client_reported_samples: number;
            /** Delivered Samples */
            delivered_samples: number;
            /** Generated Samples */
            generated_samples: number;
            /** Message Id */
            message_id: number;
            /** Playback Id */
            playback_id: string;
            /** Sample Rate */
            sample_rate: number;
            /**
             * Status
             * @enum {string}
             */
            status: "completed" | "interrupted" | "failed";
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;